finished filters
o-filter attribute, takes a name and an argument to pass to the filter. filter called on ViewModel.update, changes what's shown in the view without changing the model data. standard filter stuff;

updated ViewModel.scope object via ViewModel.watch(), now adds handlers for deep links with a reduce on the o-bind/o-update keys. Guaranteed to hit update elements who are bound to properties nested in a model object even if the model object is changed at a higher level. ex. an element bound to results.truck.view will update if results is changed or results.truck is changed, ensuring the view value is always in sync with the model