massive overhaul of event delegation and handling. Moved the delegator to an internal object Quarterback that 'calls the plays'. all OctaneBase objects have on,handle, unhandle, and fire methods. Objects register themselves with the Quarterback, who then calls delegateEvents applied to any object registered to receive those events. delegateEvents looks in an object's event map for a matching source id, and calls the appropriate array of handlers. If an object calls .handle without an eventSrc argument (2nd argument of 3), then the object will respond to ANY firing of the event.
tweaks to module loading at init, no longer fires initialize on some modules twice.

moved Template methods around, same external functionality

not re-compiling in Templates until after init. They will still get hit with the first Compiler.run before octane:ready is fired.

ViewModels hold a DOM element in memory. I don't love it, but that's the ugly truth. It's either that or querySelector for it every time the ViewModel updates.

Nesting. Do not next ViewModels in templates. they will not recompile with the rest of the directives (because they're applied with the Mediator, not the Compiler.)

fixed regexp for o-bg-image ordinance